moving "sql"-related methods into UserInputQueries
