+++ src/client/AM.PA.MacOs/PersonalAnalytics/Trackers/UserInputTracker/UserInputTracker.swift
@@ -53,22 +53,10 @@ class UserInputTracker: ITracker{
         
         inputTimer = Timer.scheduledTimer(timeInterval: inputInterval, target: self,selector: #selector(save), userInfo: nil, repeats: true)
         inputTimer?.tolerance = 5
-
-        
     }
     
     func createDatabaseTablesIfNotExist() {
-        let dbController = DatabaseController.getDatabaseController()
-        do{
-            try dbController.executeUpdate(query: "CREATE TABLE IF NOT EXISTS \(UserInputSettings.DbTableUserInput_v2) (id INTEGER PRIMARY KEY, time TEXT, tsStart TEXT, tsEnd TEXT, keyTotal INTEGER, keyOther INTEGER, keyBackspace INTEGER, keyNavigate INTEGER, clickTotal INTEGER, clickOther INTEGER, clickLeft INTEGER, clickRight INTEGER, scrollDelta INTEGER, movedDistance INTEGER)")
-            try dbController.executeUpdate(query: "CREATE TABLE IF NOT EXISTS \(UserInputSettings.DbTableKeyboard_v1) (id INTEGER PRIMARY KEY, time TEXT, timestamp TEXT, keystrokeType TEXT)")
-            try dbController.executeUpdate(query: "CREATE TABLE IF NOT EXISTS \(UserInputSettings.DbTableMouseClick_v1) (id INTEGER PRIMARY KEY, time TEXT, timestamp TEXT, x INTEGER, y INTEGER, button TEXT)")
-            try dbController.executeUpdate(query: "CREATE TABLE IF NOT EXISTS \(UserInputSettings.DbTableMouseScrolling_v1) (id INTEGER PRIMARY KEY, time TEXT, timestamp TEXT, x INTEGER, y INTEGER, scrollDelta INTEGER)")
-            try dbController.executeUpdate(query: "CREATE TABLE IF NOT EXISTS \(UserInputSettings.DbTableMouseMovement_v1) (id INTEGER PRIMARY KEY, time TEXT, timestamp TEXT, x INTEGER, y INTEGER, movedDistance INTEGER)")
-        }
-        catch{
-            print(error)
-        }
+        UserInputQueries.createDatabaseTablesIfNotExist()
     }
     
     func updateDatabaseTables(version: Int) {
@@ -119,7 +107,7 @@ class UserInputTracker: ITracker{
         (keyCount, navigateCount, deleteCount) = keystrokeController.getValues()
         keystrokeController.reset()
         tsEnd = Date()
-        dataController.saveUserInput(aggregatedInput: self)
+        UserInputQueries.saveUserInput(aggregatedInput: self)
         tsStart = Date() // reset for next aggregate
     }
